avoid a crash
How to reproduce: Open a directory in Firefox but omit the trailing
slash. Firefox will send (among others) PASV, RETR, PASV, CWD, LIST
because it cannot know that the URL points to a directory.

The second PASV connection is closed too early and therefore
fsm->datafs is NULL. This commit avoids crashing but it doesn't fix
the problem. In addition, we are setting fsm->datafs to NULL whenever
it is freed.